Add files via upload
This project involves constructing and training a Generative Adversarial Network (GAN) using PyTorch, with a focus on pre-training with the CelebA dataset and transfer learning to the AnimeFace dataset. The project covers the entire pipeline from importing libraries and downloading datasets to building the generator and discriminator models, loading pre-trained weights, and retraining the GAN on a new dataset. The CelebA dataset, containing over 200K celebrity images with annotated attributes, is used for initial training, while the AnimeFace dataset, consisting of 21,551 anime faces, serves as the downstream dataset for transfer learning. The project aims to enhance understanding of GANs, model training, and the application of transfer learning in image generation tasks.